requirements.txt
mysql-connector-python==8.1.0
pytest==7.4.0
ansible==8.6.1
docker==6.1.3